Refactor chunk-storage protocol
A chunk-storage implementation must use streams for handling data, i.e.
it expects content to be stored as a java.io.InputStream and it returns
a java.io.OutputStream when data is retrieved.

Add docstrings to the Store protocol definition and adjust the in-memory
implementation and the service endpoing accordingly.