Chunk IDs must be digest of chunk content
This is the cornerstone of having content-addressability and providing
de-duplication.

The chunk service only accepts PUTs to chunk IDs that match the digest
of the chunk content. Other requests are rejected with '400 Bad
Request'.

Currently the digest algorithm is hard-coded as SHA-256, other
implementations may be possible in the future.

[#104789932]