Store chunks in temporary file when uploading
Chunks are first stored in a temporary file, then moved to the correct
location after upload is complete. This also allows the hash to
calculated while the data is being written to disk, so the data doesn't
need to be held in memory, allowing arbitrarily large uploads (only
limited by the underlying filesystem).

[finishes #110153062]